ai_project_codex/pyproject.toml
[project]
name = "ai-project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "requests>=2.33.1",
    "pandas<3",
    "databricks-sdk>=0.46",
    "databricks-connect==17.3.*",
    "dbt-databricks>=1.10.10,<1.12",
    "mlflow[databricks]>=2.10.0",
    "scikit-learn>=1.3.0",
]
